backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.35
mysql-connector-python==8.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.1
bcrypt==4.1.3